+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Next
 
 - `Element::DomType` was renamed to `Element::DomElement`.
-- Added `runtime` module with `RuntimeSignal*` traits.
+- Added `task::TaskSignal*` traits.
 - Added `css-transpile` as an opt-in feature, as it can significatly build time.
 
 ### Fixes